IBS_CARE_AI_FINAL/backend/requirements.txt
Flask==2.2.5
gunicorn==21.2.0
python-dotenv==1.0.1
firebase-admin==6.5.0
google-generativeai==0.7.2
groq==0.4.2
httpx==0.27.2
pydantic==2.8.2
flask-cors==4.0.0
flask-mail==0.9.1
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-groq==0.2.1
langchain-core==0.3.15
langsmith==0.1.145
